Refactor internals - prepping for new Runtime impl
