+++ core/src/main/java/cucumber/runtime/CucumberBackground.java
@@ -1,7 +1,8 @@
-package cucumber.runtime.model;
+package cucumber.runtime;
 
 import gherkin.formatter.model.Background;
 
+@Deprecated
 public class CucumberBackground extends StepContainer {
     public CucumberBackground(CucumberFeature cucumberFeature, Background background) {
         super(cucumberFeature, background);
